Purpose:     Bug fix. Bug Id:     None.  Discovered during T3E port. Problem:     Fortran routines expect the ref parameter as integer type, not     integer*2 as defined. Solution:     Changed type to integer.

+++ hdf/test/tpf.f
@@ -41,7 +41,7 @@ C
       character*1 CR
       character pal1(768), pal2(768), ipal(768)
       integer ret, ref 
-      integer*2  ref1, ref2, newref1, newref2
+      integer ref1, ref2, newref1, newref2
       integer i
